Change `sotIndex` type from `Int` to `Byte` across wallet-related DTOs, mappers, and models, and refactor `presentation.ui` package imports for consistency

+++ data/user/src/main/java/com/profpay/data/user/dto/OnboardUserRequestDto.kt
@@ -50,7 +50,7 @@ data class SotAddressDto(
     @SerialName("pubKey")
     val pubKey: String,
     @SerialName("index")
-    val index: Int,
+    val index: Byte,
     @SerialName("derivationIndex")
     val derivationIndex: Int,
 )
